requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
charset-normalizer==2.1.1
dp==1.0
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.2
python-dotenv==0.21.0
python-telegram-bot==20.0a6
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
telebot==0.0.4
urllib3==1.26.13
yarl==1.8.1
